feat: Change policy gradient loss in more reliable way
Tutorial's pg loss is rely on that the 2nd action is equal zero. It is better to swap right part of the statement for generalization. Policy_grad's is also changed to rely not only on one-hot actions and it's better way to logarithmate the scalar product instead the just probabilities